Extracted configuration from ActiveMqExtension. Somewhat better separation of concerns, but more importantly, provides more freedom with respect to organising your configuration file. I.e. I can create flows without having source/sink definition in the root of my config file.
